Show all Time Entries on Homepage
Made a homepage in which a user can register and sign up and
see all time entries from the database. Data shown are currently
the seeds made in the beginning. Data is retrieved through the
use of both RESTful API of Laravel and Angular, wherein Laravel
takes care of the backend and Angular the frontend transactions.
This is, however, just the firt part of the RESTful API.
Next in line are the create, edit and delete functions.